% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example matrix data}
\format{
A numeric matrix with 2 rows and 2 columns:
\describe{
  \item{[,1]}{First column values}
  \item{[,2]}{Second column values}
}
}
\source{
Created manually for package illustration
}
\usage{
example_data
}
\description{
A simple example matrix for demonstration purposes in the package.
}
\examples{
data(example_data)
example_data

}
\keyword{datasets}
