% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{multiply_matrix_ten_times}
\alias{multiply_matrix_ten_times}
\title{Multiply two matrices and multiply 10 (using Rcpp)}
\usage{
multiply_matrix_ten_times(x, y)
}
\arguments{
\item{x}{A numeric matrix}

\item{y}{A numeric matrix}
}
\value{
A numeric matrix: 10 times the matrix product of x and y
}
\description{
This function multiplies two numeric matrices using an Rcpp implementation
and multiplies the result by 10.
}
\details{
Internally, this function uses \code{prod_cpp()}, a C++ function compiled with Rcpp,
to perform fast matrix multiplication. The result is then multiplied by a factor of 10.
This function is useful for demonstrating how to combine R and C++ in a package.
}
\examples{
mat1 <- matrix(1:6, nrow = 2)
mat2 <- matrix(1:6, nrow = 3)
multiply_matrix_ten_times(mat1, mat2)

}
\references{
See the full source code and documentation at:\cr
\url{https://github.com/downey21/r_package_practice}
}
