% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPracticePkg-package.R
\docType{package}
\name{rPracticePkg-package}
\alias{rPracticePkg-package}
\alias{rPracticePkg}
\title{rPracticePkg: Example R Package for Learning}
\description{
This package demonstrates basic R package development.
}
\details{
The goal of \code{rPracticePkg} is to provide a minimal yet complete example 
of how to build an R package from scratch. It includes:
\itemize{
  \item A simple R function (\code{add}) with roxygen2 documentation
  \item C++ integration via \code{Rcpp} (e.g., \code{t_cpp}, \code{prod_cpp})
  \item Example dataset included via \code{.rda}
  \item A vignette written in R Markdown
}

For a complete list of functions, use:
\code{library(help = "rPracticePkg")}
}
\author{
Dahun Seo
}
\keyword{package}
